requirements.txt
Flask
flask_sqlalchemy
requests
Pillow
gunicorn
